move servers into subfolder
