+++ src/jdom/jacdac-jdom.ts
@@ -32,11 +32,11 @@ export * from "./flags"
 export * from "./pack"
 export * from "./packobject"
 export * from "./light"
-export * from "./registerserver"
-export * from "./serviceserver"
-export * from "./serviceprovider"
-export * from "./serverserviceprovider"
-export * from "./controlserver"
+export * from "./servers/registerserver"
+export * from "./servers/serviceserver"
+export * from "./servers/serviceprovider"
+export * from "./servers/serverserviceprovider"
+export * from "./servers/controlserver"
 export * from "./color"
 export * from "./bridge"
 export * from "./clients/settingsclient"
