+++ src/jdom/servers/serviceserver.ts
@@ -3,21 +3,21 @@ import {
     SystemCmd,
     SystemReg,
     SystemStatusCodes,
-} from "../../jacdac-spec/dist/specconstants"
+} from "../../../jacdac-spec/dist/specconstants"
 import {
     CHANGE,
     DEVICE_CHANGE,
     PACKET_RECEIVE,
     PACKET_SEND,
     REPORT_UPDATE,
-} from "./constants"
+} from "../constants"
 import JDServerServiceProvider from "./serverserviceprovider"
-import JDEventSource from "./eventsource"
-import Packet from "./packet"
+import JDEventSource from "../eventsource"
+import Packet from "../packet"
 import JDRegisterServer from "./registerserver"
-import { isRegister, serviceSpecificationFromClassIdentifier } from "./spec"
-import { PackedValues } from "./pack"
-import JDService from "./service"
+import { isRegister, serviceSpecificationFromClassIdentifier } from "../spec"
+import { PackedValues } from "../pack"
+import JDService from "../service"
 
 const CALIBRATION_DELAY = 5000
