+++ src/servers/azureiothubhealthserver.ts
@@ -8,8 +8,8 @@ import {
     SRV_AZURE_IOT_HUB_HEALTH,
 } from "../jdom/constants"
 import Packet from "../jdom/packet"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer, { ServerOptions } from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer, { ServerOptions } from "../jdom/servers/serviceserver"
 import { delay } from "../jdom/utils"
 
 /**
