+++ src/servers/bitradioserver.ts
@@ -2,8 +2,8 @@ import { getNumber, NumberFormat, setNumber } from "../jdom/buffer"
 import { BitRadioCmd, BitRadioReg, SRV_BIT_RADIO } from "../jdom/constants"
 import { inIFrame } from "../jdom/iframeclient"
 import Packet from "../jdom/packet"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 import { memcpy, stringToBuffer } from "../jdom/utils"
 
 // keep in sync with CODAL
