+++ src/servers/capacitivebuttonserver.ts
@@ -1,6 +1,6 @@
 import { CapacitiveButtonReg, SRV_CAPACITIVE_BUTTON } from "../jdom/constants"
-import JDServiceServer, { ServerOptions } from "../jdom/serviceserver"
-import RegisterServer from "../jdom/registerserver"
+import JDServiceServer, { ServerOptions } from "../jdom/servers/serviceserver"
+import RegisterServer from "../jdom/servers/registerserver"
 
 export default class CapacitiveButtonServer extends JDServiceServer {
     readonly threshold: RegisterServer<[number]>
