+++ src/servers/characterscreenserver.ts
@@ -6,8 +6,8 @@ import {
     SRV_CHARACTER_SCREEN,
 } from "../jdom/constants"
 import Packet from "../jdom/packet"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 
 export default class CharacterScreenServer extends JDServiceServer {
     readonly message: JDRegisterServer<[string]>
