+++ src/servers/dimmerserver.ts
@@ -1,6 +1,6 @@
 import { DimmerReg, DimmerVariant, SRV_DIMMER } from "../jdom/constants"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 
 export default class DimmerServer extends JDServiceServer {
     readonly intensity: JDRegisterServer<[number]>
