+++ src/servers/hidmouseserver.ts
@@ -6,7 +6,7 @@ import {
     SRV_HID_MOUSE,
 } from "../jdom/constants"
 import Packet from "../jdom/packet"
-import JDServiceServer, { ServerOptions } from "../jdom/serviceserver"
+import JDServiceServer, { ServerOptions } from "../jdom/servers/serviceserver"
 
 /**
  * @internal
