+++ src/servers/loggerserver.ts
@@ -5,8 +5,8 @@ import {
     SRV_LOGGER,
 } from "../jdom/constants"
 import Packet from "../jdom/packet"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 
 export default class LoggerServer extends JDServiceServer {
     readonly minPriority: JDRegisterServer<[LoggerPriority]>
