+++ src/servers/matrixkeypadserver.ts
@@ -1,5 +1,5 @@
 import { MatrixKeypadReg, SRV_MATRIX_KEYPAD } from "../jdom/constants"
-import JDRegisterServer from "../jdom/registerserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
 import SensorServer from "./sensorserver"
 
 export default class MatrixKeypadServer extends SensorServer<[[number][]]> {
