+++ src/servers/motorserver.ts
@@ -1,6 +1,6 @@
 import { MotorReg, SRV_MOTOR } from "../jdom/constants"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 
 export default class MotorServer extends JDServiceServer {
     readonly duty: JDRegisterServer<[number]>
