+++ src/servers/powerserver.ts
@@ -1,6 +1,6 @@
 import { PowerReg, SRV_POWER } from "../jdom/constants"
-import JDRegisterServer from "../jdom/registerserver"
-import JDServiceServer, { ServerOptions } from "../jdom/serviceserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
+import JDServiceServer, { ServerOptions } from "../jdom/servers/serviceserver"
 
 export default class PowerServer extends JDServiceServer {
     readonly enabled: JDRegisterServer<[boolean]>
