+++ src/servers/rotaryencoderserver.ts
@@ -1,6 +1,6 @@
 import { RotaryEncoderReg, SRV_ROTARY_ENCODER } from "../jdom/constants"
 import SensorServer from "./sensorserver"
-import JDRegisterServer from "../jdom/registerserver"
+import JDRegisterServer from "../jdom/servers/registerserver"
 
 export default class RotaryEncoderServer extends SensorServer<[number]> {
     readonly clicksPerTurn: JDRegisterServer<[number]>
