+++ src/servers/settingsserver.ts
@@ -2,7 +2,7 @@ import { SettingsCmd, SettingsEvent, SRV_SETTINGS } from "../jdom/constants"
 import { jdpack } from "../jdom/pack"
 import Packet from "../jdom/packet"
 import { OutPipe } from "../jdom/pipes"
-import JDServiceServer from "../jdom/serviceserver"
+import JDServiceServer from "../jdom/servers/serviceserver"
 import { fromHex, toHex } from "../jdom/utils"
 
 export default class SettingsServer extends JDServiceServer {
